i now have block placing hooray (tis very very simple though)

+++ src/lve_util.hpp
@@ -30,7 +30,7 @@ namespace lve
 
   struct RayHit
   {
-    glm::vec3 hitPosition;
-    glm::vec3 hitDirection;
+    glm::ivec3 hitPosition;
+    glm::ivec3 hitDirection;
   };
 }
